backend/requirements.txt
fastapi
pydantic
uvicorn
bytez
python-dotenv
motor
pymongo